Add new sketch file for drawing circles
In this commit, a new sketch file (sketch_4_1.pde) is added. This file contains the code to draw two circles with diameters according to the maximum and minimum of a random dataset. The maximum value is depicted by a blue circle, and the minimum value is depicted by a red circle.